Additional changes to Raspicam
